Move styles in new folders
