+++ src/common/styles/index.js
@@ -1,0 +1,4 @@
+export { default as trackerDefaultStyle } from './trackerDefaultStyle';
+export { default as trackerDelayStyle } from './trackerDelayStyle';
+export * from './trackerDefaultStyle';
+export { default as trackerSimpleStyle } from './trackerSimpleStyle';
