+++ src/common/styles/trackerDefaultStyle.js
