+++ src/common/styles/trackerSimpleStyle.js
