+++ src/common/utils/index.js
@@ -1,7 +1,6 @@
 export { default as getMapboxStyleUrl } from './getMapboxStyleUrl';
 export { default as getMapboxMapCopyrights } from './getMapboxMapCopyrights';
 export { default as removeDuplicate } from './removeDuplicate';
-export { default as delayTrackerStyle } from './delayTrackerStyle';
-export * from './delayTrackerStyle';
-export { default as simpleTrackerStyle } from './simpleTrackerStyle';
 export * from './timeUtils';
+export * from '../styles';
+export { default as sortByDelay } from './sortByDelay';
