+++ src/mapbox/layers/TralisLayer.test.js
@@ -1,0 +1,40 @@
+import TralisLayer from './TralisLayer';
+
+let layer;
+let onClick;
+
+describe('TralisLayer', () => {
+  beforeEach(() => {
+    onClick = jest.fn();
+    layer = new TralisLayer({
+      onClick,
+    });
+  });
+
+  test('should be instanced.', () => {
+    expect(layer).toBeInstanceOf(TralisLayer);
+    expect(layer.clickCallbacks[0]).toBe(onClick);
+  });
+
+  test('#onClick', () => {
+    const f = () => {};
+    layer.onClick(f);
+    expect(layer.clickCallbacks[1]).toBe(f);
+    expect(layer.clickCallbacks.length).toBe(2);
+    layer.onClick(f);
+    expect(layer.clickCallbacks.length).toBe(2);
+  });
+
+  test('#unClick', () => {
+    const foo = () => {};
+    const bar = () => {};
+    layer.onClick(foo);
+    layer.onClick(bar);
+    expect(layer.clickCallbacks[1]).toBe(foo);
+    expect(layer.clickCallbacks[2]).toBe(bar);
+    expect(layer.clickCallbacks.length).toBe(3);
+    layer.unClick(foo);
+    expect(layer.clickCallbacks[1]).toBe(bar);
+    expect(layer.clickCallbacks.length).toBe(2);
+  });
+});
