+++ src/ol/index.js
@@ -3,10 +3,11 @@ export { default as MapboxLayer } from './layers/MapboxLayer';
 export { default as MaplibreLayer } from './layers/MaplibreLayer';
 export { default as MapboxStyleLayer } from './layers/MapboxStyleLayer';
 export { default as RoutingLayer } from './layers/RoutingLayer';
-export { default as TrackerLayer } from './layers/TrackerLayer';
 export { default as TralisLayer } from './layers/TralisLayer';
 export { default as VectorLayer } from './layers/VectorLayer';
 export { default as WMSLayer } from './layers/WMSLayer';
 export { default as CopyrightControl } from './controls/CopyrightControl';
 export { default as RoutingControl } from './controls/RoutingControl';
 export { default as StopFinderControl } from './controls/StopFinderControl';
+export * from '../common/styles';
+export * from './styles';
