+++ src/ol/styles/index.js
@@ -1,0 +1,2 @@
+export { default as fullTrajectoryStyle } from './fullTrajectoryStyle';
+export { default as fullTrajectoryDelayStyle } from './fullTrajectoryDelayStyle';
